move static methods for Text to TextWrapper
